Move try-parse-json util into separate file
