+++ workers/loc.api/helpers/utils.js
@@ -28,32 +28,6 @@ const checkParamsAuth = (args) => {
   }
 }
 
-const tryParseJSON = (
-  jsonString,
-  isNotObject
-) => {
-  try {
-    if (typeof jsonString !== 'string') {
-      return false
-    }
-
-    const obj = JSON.parse(jsonString)
-
-    if (
-      isNotObject ||
-      (
-        obj &&
-        typeof obj === 'object'
-      )
-    ) {
-      return obj
-    }
-  } catch (e) { }
-
-  return false
-}
-
 module.exports = {
-  checkParamsAuth,
-  tryParseJSON
+  checkParamsAuth
 }
