+++ workers/loc.api/helpers/utils/index.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const tryParseJSON = require('./try-parse-json')
 const collObjToArr = require('./coll-obj-to-arr')
 const getDateString = require('./get-date-string')
 const isNotSyncRequired = require('./is-not-sync-required')
@@ -12,6 +13,7 @@ const pushLargeArr = require('./push-large-arr')
 const orderBy = require('./order-by')
 
 module.exports = {
+  tryParseJSON,
   collObjToArr,
   getDateString,
   isNotSyncRequired,
