+++ workers/loc.api/helpers/utils/try-parse-json.js
@@ -1,0 +1,26 @@
+'use strict'
+
+module.exports = (
+  jsonString,
+  isNotObject
+) => {
+  try {
+    if (typeof jsonString !== 'string') {
+      return false
+    }
+
+    const obj = JSON.parse(jsonString)
+
+    if (
+      isNotObject ||
+      (
+        obj &&
+        typeof obj === 'object'
+      )
+    ) {
+      return obj
+    }
+  } catch (e) { }
+
+  return false
+}
